ignore missing files

+++ src/Terrabuild.Common/IO.fs
@@ -63,7 +63,7 @@ let copyFiles (targetDir: string) (baseDir: string) (entries: string list) =
         let target = FS.combinePath targetDir relative
         let targetDir = FS.parentDirectory target
         Directory.CreateDirectory targetDir |> ignore
-        File.Copy(entry, target, true)
+        if File.Exists(entry) then File.Copy(entry, target, true)
     if entries |> List.isEmpty then None
     else Some targetDir
